+++ src/frequency-histogram/pure_lfu_histogram.cpp
@@ -2,7 +2,7 @@
 #include <string>
 #include "frequency-histogram/pure_lfu_histogram.h"
 
-std::uint64_t PureLfuHistogram::Add(const std::string &query) {
+std::uint32_t PureLfuHistogram::Add(const std::string &query) {
   if (histogram.find(query) != histogram.end()) {
     histogram[query]++;
   } else {
@@ -12,6 +12,6 @@ std::uint64_t PureLfuHistogram::Add(const std::string &query) {
   return histogram[query];
 }
 
-std::uint64_t PureLfuHistogram::Estimate(const std::string &query) const {
+std::uint32_t PureLfuHistogram::Estimate(const std::string &query) const {
   return histogram.at(query);
 }
